src/requirements.txt
Django==1.10.4
django-choices==1.4.4
django-polymorphic==1.0.2
django-relativefilepathfield==1.0.3
mysqlclient==1.3.9
six==1.10.0
python-telegram-bot